Move get_query function to team_tags file
