+++ di_website/common/constants.py
@@ -51,5 +51,3 @@ PODCAST_PROVIDERS = [
     ('google', 'Google'),
     ('amazon', 'Amazon'),
 ]
-
-AUTHOR_INDEXES = [0,1,2,3,4]
